requirements.txt
python-telegram-bot>=21.4
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
anthropic>=0.39.0
httpx>=0.27.0
